feat: 리다이렉트 페이지를 사용자 친화적으로 변경
