+++ apps/web/src/app/App.tsx
@@ -1,8 +1,9 @@
 import { HashRouter, Navigate, Route, Routes, useLocation } from "react-router-dom";
 import { GoogleReCaptchaProvider } from "react-google-recaptcha-v3";
-import { SignInCompletePage, SignInPage, SignUpPage } from "@/pages/auth";
+import { SignInPage, SignUpPage } from "@/pages/auth";
 import { AuthLayout } from "@/app/layouts/auth";
 import { OnboardingPage } from "@/pages/onboarding";
+import { RedirectPage } from "@/pages/redirect";
 
 const RootRoute = () => {
   const location = useLocation();
@@ -28,8 +29,8 @@ function App() {
         <Routes>
           <Route element={<AuthLayout />}>
             <Route path="/" element={<RootRoute />} />
+            <Route path="/redirect" element={<RedirectPage />} />
             <Route path="/sign-in" element={<SignInPage />} />
-            <Route path="/sign-in-complete" element={<SignInCompletePage />} />
             <Route path="/sign-up" element={<SignUpPage />} />
             <Route path="*" element={<SignInRedirect />} />
           </Route>
