+++ apps/web/src/pages/auth/index.ts
@@ -1,3 +1,2 @@
 export { SignInPage } from "./sign-in/SignIn";
-export { SignInCompletePage } from "./sign-in-complete/SignInComplete";
 export { SignUpPage } from "./sign-up/SignUp";
