+++ apps/web/src/pages/redirect/index.ts
@@ -1,0 +1,1 @@
+export { RedirectPage } from "./Redirect.tsx";
